Major Updates
Pthonious Main scripts runs all other scripts on seperate instances allowing multiple progams to be run side by side easily.
The Hash.py script can create hashes of any written text and can be ran through 8 different commonly used algorithms.
The HashMatcher.py script takes in a hash then autodetects the types of hash it can be and then gives options such as the length of the password and types of characters in it and the ammount of threads that will deal with the process at once. It's been tested that just using 1 thread produces the best results since the program constantly outputs to the counter variable that will keep count of attemps, but since the program is for demonstrative purposes that is fine.

Side Notes:: Add more character options to the HashMatcher such as only special characters or number with special characters.
Also fix all exits on the code and instead allow the user to re-run the program or the specific action in code by turning them into functions(def in python).